Moved common functions to script_utils.py
